+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 1.7 (upcoming)
+* Revised: Minor improvement for procceding with semiblind ("file-based") technique, once the user provides the path of web server's root directory.
 * Fixed: Minor fix regarding the lack of http/s to the user-defined URL(s).
 * Added: New option "--skip-empty" for skipping the testing of the parameter(s) with empty value(s).
 * Fixed: Improvement regarding testing the parameter(s) with empty value(s).